requirements2.txt
Adafruit-Blinka==4.1.0
adafruit-circuitpython-adxl34x==1.10.4
adafruit-circuitpython-busdevice==4.2.1
adafruit-circuitpython-gps==3.5.2
adafruit-circuitpython-max31865==2.1.4
Adafruit-PlatformDetect==2.3.0
Adafruit-PureIO==1.0.4
astroid==2.3.3
autopep8==1.5
dj-database-url==0.5.0
# Django==2.2.10
# django-webpack-loader==0.7.0
# djangorestframework==3.11.0
# gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pigpio==1.45
psycopg2==2.8.4
pycodestyle==2.5.0
pyftdi==0.44.2
pylint==2.4.4
pynmea2==1.15.0
pyserial==3.4
pytz==2019.3
pyusb==1.0.2
quick2wire-api==0.0.0.2
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
six==1.14.0
spidev==3.4
sqlparse==0.3.0
sysv-ipc==1.0.1
typed-ast==1.4.1
# whitenoise==5.0.1
wrapt==1.11.2
